Seperating out the eval so it can easily be run outside of quantization
