+++ recipes/alpamayo1_5_quant/README.md
@@ -4,11 +4,11 @@ This Recipe defines a reproducible post-training quantization (PTQ) procedure fo
 
 ## Prerequisites
 
-This recipe is tested in the following settings. Other settings may also work but not guaranteed.
+This recipe is tested in the following settings.
 
 - NVIDIA B300 GPU with CUDA 13
 - Python 3.12
-- Python Libraries: torch==2.12.0, torchvision==0.27.0, nvidia-modelopt==0.43.0
+- Python Libraries: torch==2.8.0, torchvision==0.27.0, nvidia-modelopt==0.43.0
 
 **NVIDIA Model Optimizer (ModelOpt)** is a library comprising state-of-the-art model optimization techniques including quantization and sparsity to compress models. In this recipe, we utilize ModelOpt to quantize Alpamayo 1.5.
 
@@ -22,9 +22,12 @@ This recipe is tested in the following settings. Other settings may also work bu
     1. [Settings](#settings)
     2. [FP8 Quantization](#fp8-quantization)
     3. [Autoquant (NVFP4 + FP8 Mixed Precision)](#autoquant)
-    4. [Expected Output](#expected-runtime-behaviors-and-outputs)
+3. [Evaluation](#evaluation)
+    1. [Settings](#evaluation-settings)
+    2. [Running evaluation](#running-evaluation)
+    3. [Expected output](#expected-output)
 
-<!-- 3. [FAQ](#faq) -->
+<!-- 4. [FAQ](#faq) -->
 
 ## Getting started
 
@@ -57,6 +60,7 @@ Set the following once per session (or add to `~/.bashrc`):
 export ALPAMAYO_WORKSPACE="$YOUR_HOME/alpamayo-recipes"
 export ALPAMAYO_MODEL_DIR="$YOUR_HOME/alpamayo_model_converted_from_hf"
 export ALPAMAYO_PAI_LOCAL_DIR="$YOUR_HOME/PAI_mini"
+export ALPAMAYO_LOG_DIR="$YOUR_HOME/alpamayo_logs"
 
 # ── Cache ────────────────────────────────────────────────────────
 export HF_HOME="$YOUR_HOME/.cache/huggingface"
@@ -91,47 +95,83 @@ hf auth login
 
 ## Quantization
 
+Quantization (`quantize.py`) calibrates and saves a quantized checkpoint. Evaluation (`eval.py`) is a separate step that loads any checkpoint — base or quantized — and reports minADE metrics.
+
 ### Settings
 
 The quantization path is controlled by the following arguments:
 
-- `--parquet <path>`: evaluation clip source. `1005_7cam_gold_eval_metadb_public.parquet` is used by default.
-- `--quant_format fp8`: enables FP8 PTQ.
-- `--quant_algo <algo>`: keeps default `max` for FP8 runs.
-- `--quant_weight_only` (optional): enables FP8 weight-only PTQ.
-- `--calib_parquet <path>`: calibration clip source. `0417_5k_train_set_for_calibration_25.10.parquet` is used by default.
-- `--num_of_calib_clips <N>`: number of calibration clips (1 to 5000). `100` is used by default.
-
-Please refer to `quantize.py` to find the usage of more arguments.
+- `--ckpt <path_or_hub_id>`: model to quantize. Defaults to `nvidia/Alpamayo-1.5-10B`.
+- `--quant_format <fmt>`: quantization format — `fp8`, `nvfp4`, `w4a8_nvfp4_fp8`, or `auto`. **Required.**
+- `--quant_algo <algo>`: calibration algorithm. Defaults to `max`; `smoothquant` is also supported.
+- `--quant_weight_only` (optional): enables weight-only quantization.
+- `--auto_quantize_bits <N>`: effective-bits budget for AutoQuant. Defaults to `4.8`.
+- `--calib_parquet <path>`: calibration clip source. Defaults to `0417_5k_train_set_for_calibration_25.10.parquet`.
+- `--num_of_calib_clips <N>`: number of calibration clips (1–5000). Defaults to `100`.
+- `--save_model_dir <path>`: directory under which the quantized model is saved. **Required.** The checkpoint is written to a subdirectory named after the format and calibration settings (e.g. `alpamayo1.5_fp8_calib100`).
 
 ### FP8 quantization
 
-Run an example command below to quantize alpamayo1.5 in FP8 and save the quantized model:
+Quantize and save a FP8 checkpoint:
 
 ```bash
 uv run --active quantize.py --quant_format=fp8 --num_of_calib_clips=100 --save_model_dir=./outputs
 ```
 
+To run in the background and capture logs:
+
+```bash
+nohup uv run --active quantize.py --quant_format=fp8 --num_of_calib_clips=100 --save_model_dir=./outputs > quantize_fp8.log 2>&1 &
+```
+
 ### AutoQuant
 
-Autoquant is a tool that allows for mixed NVFP4 + FP8 quantization while still remaining lossless.
+AutoQuant searches per-layer across NVFP4 and FP8 under an effective-bits budget, enabling mixed-precision quantization with minimal accuracy loss.
 
-Run an example command below to quantize alpamayo1.5 in AutoQuant (FP8 + NVFP4) with 6.5 effective bits and save the quantized model:
+Quantize and save an AutoQuant (FP8 + NVFP4) checkpoint at 6.5 effective bits:
 
 ```bash
 uv run --active quantize.py --quant_format=auto --auto_quantize_bits=6.5 --num_of_calib_clips=100 --save_model_dir=./outputs
 ```
 
-### Expected runtime behaviors and outputs
+## Evaluation
 
-During a correct run, logs would show:
+### Evaluation settings
 
-- Calibration clips are loaded from `--calib_parquet`.
-- Calibration loop progress (`calibration: ...%`) is executed.
-- Quantization summary is printed.
-- Evaluation starts and reports per-clip metrics and final averages.
+Evaluation (`eval.py`) loads any Alpamayo 1.5 checkpoint — base FP16 or a quantized checkpoint saved by `quantize.py` — and reports per-clip and average minADE metrics.
 
-After the evaluation, you will see the following outputs:
+- `--ckpt <path_or_hub_id>`: checkpoint to evaluate. Defaults to `nvidia/Alpamayo-1.5-10B`. Pass the path saved by `quantize.py` to evaluate a quantized model.
+- `--parquet <path>`: evaluation clip source. Defaults to `1005_7cam_gold_eval_metadb_public.parquet`.
+- `--limit <N>`: number of clips to evaluate. Defaults to `644`.
+- `--num_traj_samples <N>`: trajectory samples per clip. Defaults to `6`.
+- `--print_every <N>`: log a progress line every N clips. Defaults to `25`.
+- `--seed <N>`: random seed for reproducible sampling. Set to `-1` to disable. Defaults to `42`.
 
-- Average minADE
-- Average evaluation time per clip
+### Running evaluation
+
+Evaluate the base FP16 model:
+
+```bash
+uv run --active eval.py
+```
+
+Evaluate a quantized checkpoint saved by `quantize.py`:
+
+```bash
+uv run --active eval.py --ckpt ./outputs/alpamayo1.5_fp8_calib100
+```
+
+To run in the background and capture logs:
+
+```bash
+nohup uv run --active eval.py --ckpt ./outputs/alpamayo1.5_fp8_calib100 > eval_fp8.log 2>&1 &
+```
+
+### Expected output
+
+During a correct run, logs will show:
+
+- Clip IDs loaded from `--parquet`.
+- Evaluation progress (`Evaluating clips: ...%`) with per-clip minADE and timing every `--print_every` clips.
+- Any failed clips and their errors.
+- Final summary: average minADE and average evaluation time per clip.
