+++ recipes/alpamayo1_5_quant/pyproject.toml
@@ -14,8 +14,8 @@ dependencies = [
   "pandas>=2.3.3",
   "physical-ai-av==0.2.0",
   "pillow>=12.0.0",
-  "torch==2.12.0", #switch back to 2.8 later
-  "torchvision==0.27.0",
+  "torch==2.8.0",
+  "torchvision==0.23.0",
   "transformers==4.57.1",
   "flash-attn>=2.8.3",
   "matplotlib>=3.10.7",
@@ -45,9 +45,6 @@ dev = [
 
 [tool.uv]
 no-build-isolation-package = ["flash-attn"]
-override-dependencies = [
-    "torch==2.12.0", #switch back to 2.8 later
-]
 
 [tool.uv.sources]
 alpamayo1_5 = { git = "https://github.com/NVlabs/alpamayo1.5.git" }
